+++ src/utils/editOntology/setOntologyRestoreEdge.js
@@ -6,8 +6,6 @@ import { POST_CREATE_EDGE } from '../../constants/api'
 import httpCall from '../apiCalls/httpCall'
 import showNotification from '../notifications/showNotification'
 import { NOTIFY_SUCCESS, NOTIFY_WARNING } from '../../constants/notifications'
-import countNodes from '../nodesEdgesUtils/countNodes'
-import countEdges from '../nodesEdgesUtils/countEdges'
 
 /**
  * ADd ontology edge
@@ -113,7 +111,7 @@ const setOntologyRestoreEdge = async ({
       if (
         isFromVisible
         && isToVisible) {
-        addEdge(edge)
+        addEdge({ edge, addNumber })
 
         // add connections
         if (!newNodesEdges[from].includes(id)) {
@@ -152,9 +150,6 @@ const setOntologyRestoreEdge = async ({
       message,
       type: NOTIFY_SUCCESS
     })
-
-    setStoreState('availableNodesCount', countNodes())
-    setStoreState('availableEdgesCount', countEdges())
   }
 }