+++ src/utils/nodesEdgesUtils/addNode.js
@@ -2,18 +2,21 @@ import store from '../../store'
 
 /**
  * Add node to graph
- * @param  {Object}   nodeObject    Node object with at least id and label keys
+ * @param  {Object}     params
+ * @param  {Object}     params.node           Node object with at least id and label keys
+ * @param  {function}   params.addNumber      Add number action
  * @return { undefined }
 \ */
-const addNode = (nodeObject) => {
+const addNode = ({ node, addNumber }) => {
   const {
     availableNodes
   } = store.getState()
 
-  const isNodeNotVisible = availableNodes.get(nodeObject.id) === null
+  const isNodeNotVisible = availableNodes.get(node.id) === null
 
   if (isNodeNotVisible) {
-    availableNodes.add(nodeObject)
+    availableNodes.add(node)
+    addNumber('availableNodesCount', 1)
   }
 }